GeoscapeState::time1Day() - do permil chance to die in sickbay
FundingState: change Ok to Cancel
	- varnames
RuleArmor::load() - remove obsolete identifiers
	- cast DAMAGE_TYPES to size_t

MoreCivilians.rul - add battleType=11

+++ src/Ruleset/RuleArmor.h
@@ -49,7 +49,7 @@ class RuleArmor
 {
 
 	public:
-		static const int DAMAGE_TYPES = 10;
+		static const size_t DAMAGE_TYPES = 10;
 
 
 private:
